refactor: separate telegram interface
