+++ README.md
@@ -91,7 +91,7 @@ npm run lint:fix
 ## Структура проекта
 
 - `src/index.ts` — точка входа приложения.
-- `src/infrastructure/telegram/TelegramBot.ts` — логика бота и обработка сообщений.
+- `src/view/telegram/TelegramBot.ts` — логика бота и обработка сообщений.
 - `src/infrastructure/external/ChatGPTService.ts` — взаимодействие c OpenAI.
 - `src/application/use-cases/chat/ChatMemory.ts` — хранение истории и её суммаризация.
 - `dist/` — собранные файлы.
