+++ src/view/telegram/triggers/NameTrigger.ts
@@ -4,17 +4,17 @@ import type { Context } from 'telegraf';
 import {
   ENV_SERVICE_ID,
   type EnvService,
-} from '../application/interfaces/env/EnvService.interface';
-import type { Logger } from '../application/interfaces/logging/Logger.interface';
+} from '../../../application/interfaces/env/EnvService.interface';
+import type { Logger } from '../../../application/interfaces/logging/Logger.interface';
 import {
   LOGGER_FACTORY_ID,
   type LoggerFactory,
-} from '../application/interfaces/logging/LoggerFactory.interface';
+} from '../../../application/interfaces/logging/LoggerFactory.interface';
 import type {
   Trigger,
   TriggerContext,
   TriggerResult,
-} from '../domain/triggers/Trigger.interface';
+} from '../../../domain/triggers/Trigger.interface';
 
 @injectable()
 export class NameTrigger implements Trigger {
