+++ src/view/telegram/windowConfig.ts
