+++ test/InterestTrigger.test.ts
@@ -4,7 +4,7 @@ import { describe, expect, it, vi } from 'vitest';
 import { DefaultDialogueManager } from '../src/application/use-cases/chat/DefaultDialogueManager';
 import type { DialogueManager } from '../src/application/interfaces/chat/DialogueManager.interface';
 import { InterestChecker } from '../src/application/interfaces/interest/InterestChecker.interface';
-import { InterestTrigger } from '../src/triggers/InterestTrigger';
+import { InterestTrigger } from '../src/view/telegram/triggers/InterestTrigger';
 import { TriggerContext } from '../src/domain/triggers/Trigger.interface';
 import type { LoggerFactory } from '../src/application/interfaces/logging/LoggerFactory.interface';
