+++ test/Triggers.test.ts
@@ -3,9 +3,9 @@ import { describe, expect, it, vi } from 'vitest';
 
 import { DefaultDialogueManager } from '../src/application/use-cases/chat/DefaultDialogueManager';
 import { TestEnvService } from '../src/infrastructure/config/TestEnvService';
-import { MentionTrigger } from '../src/triggers/MentionTrigger';
-import { NameTrigger } from '../src/triggers/NameTrigger';
-import { ReplyTrigger } from '../src/triggers/ReplyTrigger';
+import { MentionTrigger } from '../src/view/telegram/triggers/MentionTrigger';
+import { NameTrigger } from '../src/view/telegram/triggers/NameTrigger';
+import { ReplyTrigger } from '../src/view/telegram/triggers/ReplyTrigger';
 import { TriggerContext } from '../src/domain/triggers/Trigger.interface';
 import type { LoggerFactory } from '../src/application/interfaces/logging/LoggerFactory.interface';
 import type { EnvService } from '../src/application/interfaces/env/EnvService.interface';
