+++ vitest.config.ts
@@ -13,8 +13,7 @@ export default defineConfig({
         'src/container.ts',
         'src/infrastructure/persistence/**',
         'src/infrastructure/config/**',
-        'src/infrastructure/telegram/**',
-        'src/triggers/**',
+        'src/view/telegram/**',
         'src/domain/entities/**',
         '**/*.interface.ts',
       ],
